src/backend/user_service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
alembic==1.13.1
cffi==1.16.0
click==8.1.7
fastapi==0.110.0
grpcio==1.62.1
grpcio-tools==1.62.1
gunicorn==21.2.0
h11==0.14.0
httpx==0.27.0
idna==3.6
packaging==24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
python-dotenv==1.0.1
pytest==8.1.1
pytest-asyncio==0.23.6
redis==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
